Improve Riemann server connection resiliency
This change improves the way the Riemann server connection is
established: instead of trying to connect only once during startup, use
the same method as for the database handle i.e. test the handle at every
check loop call and try to recreate the connection if it has been found
broken. This allows for a recovery in the case of a temporary server
unavailability.